Redesign rendering
Add a stateless struct `ProjectData` to represent all tasks. Right now,
it's only used to initialize the view. Later, this will be the struct to
be serialised and to store the tasks for the app.

Add a struct `ProjectDataEditView` that wraps around an instance of
`ProjectData` and contains all the state for the user interaction and
rendering in the terminal. It uses a trait `TaskView` to model all
different interactions with the terminal. This will be useful for
creating a dialogue to create new tasks

The ability to create new tasks is not yet reimplemented. Right now,
the only interaction possible is to navigate through dummy tasks and to
mark them as done. I will add new interactions once I layed some more
groundwork (like implementing the window for command previews).