WIP Add action preview
Removed `tick` method from app, since I don't really need it and it
basically duplicated handling of AppEvent.
Added a new event to trigger repopulation of tasks. This is a bit hacky
since a better place to handle this is in the app directly, but doing it
like this creates an easier control flow.

Add struct `ActionPreviewSet` to store all available actions. Pass it
through the `ProjectDataEditView` to fill available actions recursively.